Break up EI test class
