+++ spec/suitec/whiteboard_assets_spec.rb
@@ -21,7 +21,7 @@ describe 'Whiteboard Add Asset', order: :defined do
     @canvas = Page::CanvasPage.new @driver
     @cal_net = Page::CalNetPage.new @driver
     @asset_library = Page::SuiteCPages::AssetLibraryDetailPage.new @driver
-    @engagement_index = Page::SuiteCPages::EngagementIndexPage.new @driver
+    @engagement_index = Page::SuiteCPages::EngagementIndexConfigPage.new @driver
     @whiteboards = Page::SuiteCPages::WhiteboardsPage.new @driver
 
     # Create test course
