+++ util/spec_helper.rb
@@ -80,5 +80,6 @@ require_relative '../pages/suitec/asset_library_list_view_page'
 require_relative '../pages/suitec/asset_library_detail_page'
 require_relative '../pages/suitec/asset_library_manage_assets_page'
 require_relative '../pages/suitec/engagement_index_page'
+require_relative '../pages/suitec/engagement_index_config_page'
 require_relative '../pages/suitec/whiteboards_page'
 require_relative '../pages/suitec/impact_studio_page'
